rename cards component to MainCard
